Can now query the games and fish in games
