--- handlers/FishTracker.js
@@ -1,24 +1,0 @@
-const { db } = require("../utils/admin");
-
-exports.fishTracker = async (req, res) => {
-    const fishTrackerRef = db.collection('CrimsCastFishTacker');
-    try{
-            // fishTrackerRef.doc(game).collection('Fish').get().then((snapshot) => {console.log(snapshot.docs.map((doc) => ({
-            //     id: doc.id,
-            //     ...doc.data(),
-            // }))
-            // )});
-            fishTrackerRef.get().then((snapshot) => {
-            const data = snapshot.docs.map((doc) => ({
-            id: doc.id,
-            ...doc.data(),
-        }));
-            console.log(data);
-            return res.status(201).json(data);
-        })
-    } catch (error) {
-        return res
-        .status(500)
-        .json({ general: "Something went wrong, please try again"});          
-    }
-};
