+++ index.js
@@ -2,7 +2,8 @@
 const express = require('express');
 const app = express();
 
-const { fishTracker } = require('./handlers/FishTracker');
+const { gameList } = require('./handlers/GameList');
+const { gamesFish } = require('./handlers/GamesFish');
 const { addFish } = require('./handlers/AddFish');
 
 const PORT = process.env.PORT || 5050
@@ -14,8 +15,9 @@ app.get('/', (req, res) => {
     res.send('This is my demo project')
 })
 
-app.get('/fish', fishTracker);
+app.get('/gamelist', gameList);
 
+app.get('/gamesfish', gamesFish);
 
 app.post('/newfish', addFish);
